blockcmd: Widen scsilun2u64 shifts to u64 to avoid undefined behaviour
scsilun2u64 reassembles the four big-endian 16-bit words of
a SCSI LUN into a 64-bit value:

    ret |= be16_to_cpu(scsi_lun->lun[i]) << (16 * i);

be16_to_cpu returns u16. The usual integer promotions
promote that to int (32-bit on x86), so for i = 2 and i = 3
the left operand of << is int and the shift count is 32 or
48. C99 6.5.7 paragraph 3 makes a shift by a count greater
than or equal to the operand width undefined behaviour. On
x86 the SHL instruction masks the shift count to five bits,
so << 32 wraps to << 0 and << 48 wraps to << 16, both
feeding back into bits 0-31 of ret. The high 32 bits of the
assembled LUN are therefore always zero, which makes the

    if (lun >> 32) continue;

skip-test in scsi_rep_luns_scan inert: a hierarchical LUN
whose high words are non-zero gets silently truncated to a
LUN that fits in 32 bits, and that truncated value may match
a different (already-existing) LUN on the same target.

Cast the be16_to_cpu result to u64 before shifting so the
shift is well-defined and the high half of the LUN survives.

+++ src/hw/blockcmd.c
@@ -236,7 +236,7 @@ static u64 scsilun2u64(struct scsi_lun *scsi_lun)
     int i;
     u64 ret = 0;
     for (i = 0; i < ARRAY_SIZE(scsi_lun->lun); i++)
-        ret |= be16_to_cpu(scsi_lun->lun[i]) << (16 * i);
+        ret |= (u64)be16_to_cpu(scsi_lun->lun[i]) << (16 * i);
     return ret;
 }
